+++ package.json
@@ -13,6 +13,7 @@
 		"@testing-library/user-event": "^13.2.1",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
+		"react-loading": "^2.0.3",
 		"react-router-dom": "^6.2.1",
 		"react-scripts": "5.0.0",
 		"sass": "^1.45.1",
